+++ src/parser/group.go
@@ -1,0 +1,15 @@
+package parser
+
+import (
+	"github.com/sevenreup/chewa/src/ast"
+	"github.com/sevenreup/chewa/src/token"
+)
+
+func (p *Parser) parseGroupedExpression() ast.Expression {
+	p.nextToken()
+	exp := p.parseExpression(LOWEST)
+	if !p.expectPeek(token.CLOSING_PAREN) {
+		return nil
+	}
+	return exp
+}
