+++ src/parser/number.go
@@ -1,0 +1,19 @@
+package parser
+
+import (
+	"fmt"
+	"github.com/sevenreup/chewa/src/ast"
+	"strconv"
+)
+
+func (p *Parser) parseIntegerLiteral() ast.Expression {
+	lit := &ast.IntegerLiteral{Token: p.curToken}
+	value, err := strconv.ParseInt(p.curToken.Literal, 0, 64)
+	if err != nil {
+		msg := fmt.Sprintf("could not parse %q as integer", p.curToken.Literal)
+		p.errors = append(p.errors, msg)
+		return nil
+	}
+	lit.Value = value
+	return lit
+}
