+++ src/parser/string.go
@@ -1,0 +1,7 @@
+package parser
+
+import "github.com/sevenreup/chewa/src/ast"
+
+func (p *Parser) parseStringLiteral() ast.Expression {
+	return &ast.StringLiteral{Token: p.curToken, Value: p.curToken.Literal}
+}
